PROD-707 Embiggen PK of tables related to call caching  (#6815)

+++ database/sql/src/main/scala/cromwell/database/slick/tables/CallCachingDetritusEntryComponent.scala
@@ -12,13 +12,13 @@ trait CallCachingDetritusEntryComponent {
 
   class CallCachingDetritusEntries(tag: Tag)
     extends Table[CallCachingDetritusEntry](tag, "CALL_CACHING_DETRITUS_ENTRY") {
-    def callCachingDetritusEntryId = column[Int]("CALL_CACHING_DETRITUS_ENTRY_ID", O.PrimaryKey, O.AutoInc)
+    def callCachingDetritusEntryId = column[Long]("CALL_CACHING_DETRITUS_ENTRY_ID", O.PrimaryKey, O.AutoInc)
 
     def detritusKey = column[String]("DETRITUS_KEY", O.Length(255))
 
     def detritusValue = column[Option[SerialClob]]("DETRITUS_VALUE")
 
-    def callCachingEntryId = column[Int]("CALL_CACHING_ENTRY_ID")
+    def callCachingEntryId = column[Long]("CALL_CACHING_ENTRY_ID")
 
     override def * = (detritusKey, detritusValue, callCachingEntryId.?, callCachingDetritusEntryId.?) <>
       (CallCachingDetritusEntry.tupled, CallCachingDetritusEntry.unapply)
@@ -37,7 +37,7 @@ trait CallCachingDetritusEntryComponent {
     callCachingDetritusEntries returning callCachingDetritusEntries.map(_.callCachingDetritusEntryId)
 
   val callCachingDetritusEntriesForCallCachingEntryId = Compiled(
-    (callCachingEntryId: Rep[Int]) => for {
+    (callCachingEntryId: Rep[Long]) => for {
       callCachingDetritusEntry <- callCachingDetritusEntries
       if callCachingDetritusEntry.callCachingEntryId === callCachingEntryId
     } yield callCachingDetritusEntry
