+++ supportedBackends/google/pipelines/common/src/test/scala/cromwell/backend/google/pipelines/common/callcaching/PipelinesApiBackendCacheHitCopyingActorSpec.scala
@@ -518,7 +518,7 @@ class PipelinesApiBackendCacheHitCopyingActorSpec extends TestKitSuite
     actorUnderTest
   }
 
-  private def buildCopyCommand(hitId: Int, bucket: String): CopyOutputsCommand = {
+  private def buildCopyCommand(hitId: Long, bucket: String): CopyOutputsCommand = {
     val callRoot = s"gs://$bucket/workflow-id/call-name"
     val rcFile = callRoot + "/rc"
